boardstar/requirements.txt
streamlit
langchain
google-cloud-aiplatform
faiss-cpu
sentence-transformers
python-dotenv